% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_col_positions.R
\name{as_col_positions}
\alias{as_col_positions}
\alias{as_col_positions.table_design_fwf}
\title{Convert table_design_fwf to col_positions for readr}
\usage{
as_col_positions(x)

\method{as_col_positions}{table_design_fwf}(x)
}
\arguments{
\item{x}{a \code{table_design_fwf} object.}
}
\value{
a \code{list} with columns \code{start}, \code{end}, \code{col_name}.

a \code{list} with columns \code{start}, \code{end}, \code{col_name}
}
\description{
Convert \code{\link{table_design_fwf}} objects to lists that can be used as
\code{col_positions} argument to \code{\link[readr:read_fwf]{readr::read_fwf()}}
}
