% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_sql.R
\name{as_sql}
\alias{as_sql}
\alias{as_sql.table_design_sql}
\title{Generate an SQL CREATE TABLE Statement From a tabde Table Design}
\usage{
as_sql(x, ...)

\method{as_sql}{table_design_sql}(x, tname, ...)
}
\arguments{
\item{x}{any \R object}

\item{...}{passes on to methods}

\item{tname}{\code{character} scalar. name of target table}
}
\value{
\code{character} scalar. An SQL CREATE TABLE statement.
}
\description{
This does not perform any sanity check on the input data. \code{col_name} and
\code{sql_type} must be compatible with the DBMS that you want to use.
}
\examples{

td <- tabde_sql(
  col_name = c("numbers", "letters"),
  col_type = c("integer", "character"),
  sql_type = c("smallint", "varchar(255)")
)

as_sql(td, "test_table")

}
