% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabde_constraints.R
\name{as_sql_header}
\alias{as_sql_header}
\alias{as_sql_header.list}
\alias{as_sql_header.data.frame}
\title{Coerce to sql_header}
\usage{
as_sql_header(x)

\method{as_sql_header}{list}(x)

\method{as_sql_header}{data.frame}(x)
}
\arguments{
\item{x}{any supported \R object}
}
\value{
a \code{sql_header} object
}
\description{
Coerce to sql_header
}
