% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabde_constraints.R
\name{as_tabde_constraints.data.frame}
\alias{as_tabde_constraints.data.frame}
\title{Title}
\usage{
\method{as_tabde_constraints}{data.frame}(x)
}
\arguments{
\item{x}{}
}
\value{

}
\description{
Title
}
