% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabde_constraints.R
\name{as_tabde_header}
\alias{as_tabde_header}
\title{Coerce to tabde_header}
\usage{
as_tabde_header(x)
}
\arguments{
\item{x}{any supported \R object}
}
\value{
a \code{tabde_constraints} object
}
\description{
Coerce to tabde_header
}
