% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_design.R
\name{as_table_design}
\alias{as_table_design}
\alias{as_table_design.data.frame}
\alias{as_table_design_fwf}
\alias{as_table_design_sql}
\alias{is_table_design}
\alias{is_table_design_sql}
\alias{is_table_design_fwf}
\title{Coerce object to table_design}
\usage{
as_table_design(x, ...)

\method{as_table_design}{data.frame}(x, ...)

as_table_design_fwf(x)

as_table_design_sql(x, constraints = attr(x, "constraints"))

is_table_design(x)

is_table_design_sql(x)

is_table_design_fwf(x)
}
\arguments{
\item{x}{any \R object}

\item{...}{passed on to methods}

\item{constraints}{a \link{tabde_constraints} object}
}
\value{
a \code{table_design} object of the appropriate subclass
}
\description{
Coerce object to table_design
}
