% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_typecast.R
\name{df_typecast}
\alias{df_typecast}
\alias{df_typecast_list}
\title{Typecast columns of a data.frame}
\usage{
df_typecast(x, table_design, silent = FALSE)

df_typecast_list(x, conv, silent)
}
\arguments{
\item{x}{a \code{data.frame} or \code{list}}

\item{table_design}{a \link{table_design}}

\item{silent}{\code{logical}. Should warnings be displayed?}

\item{conv}{a list of the form list(colname = "coltype")}
}
\value{
Either a \code{data.frame} or \code{list} with typecasted columns/elements
(depending on the type of \code{x})
}
\description{
Bulk-typecast columns of a \code{data.frame}. Use with care, will introduce NAs
for impossible casts (like the base \R as.* functions)
}
\details{
\code{df_typecast()} uses a \link{table_design} Object to specify column types, while
\code{df_typecast_list()} uses a \code{list} (see examples).
}
\examples{


dat <- data.frame(
  foo = c('5', '6', '5'),
  bar = factor(c('a', 'b', 'c')),
  stringsAsFactors = FALSE
)
str(dat)


td <- tabde(
  col_name = c("foo", "bar"),
  col_type = c("character", "factor")
)
res <- df_typecast(dat, td)
str(res)


res <- df_typecast_list(
  dat,
  list(foo = "character", bar = "factor")
)
str(res)



}
