% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_sql.R
\name{generate_sql_create_table}
\alias{generate_sql_create_table}
\title{Generate SQL CREATE TABLE statements}
\usage{
generate_sql_create_table(tname, col_names, col_types, sql_opts = rep("",
  length(col_names)))
}
\arguments{
\item{tname}{name of target sql table}

\item{col_names}{column names of target sql table}

\item{col_types}{column types of target sql table. Columns of type NA will
be skipped}

\item{sql_opts}{column options of target sql table (for example \code{NOT NULL})}
}
\value{
a \code{CREATE TABLE} statement as a text string
}
\description{
Creates SQL CREATE TABLE statements from a vector of column names and
a vector of column types
}
\examples{

generate_sql_create_table(
  "example.table",
  c("numbers", "animals"),
  c("integer", "varchar(8)"),
  c("NOT NULL", "")
)


}
