% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tabde.R
\name{get_tabde}
\alias{get_tabde}
\alias{get_tabde_sql}
\alias{get_tabde_fwf}
\title{Get table description from data.frame}
\usage{
get_tabde(x)

get_tabde_sql(x)

get_tabde_fwf(x)
}
\arguments{
\item{x}{a \code{data.frame}}
}
\value{
a \code{data.frame} with class \link{table_design}
}
\description{
\code{get_tabde()} generates a table design from a \code{data.frame}. \code{get_tabde_sql()}
and \code{get_tabde_fwf()} do the same, but also add the respective dummy columns
for \code{table_design_sql} and \code{table_design_fwf} Objects.
}
\examples{

get_tabde(iris)
get_tabde_sql(iris)
get_tabde_fwf(iris)
}
