% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_design.R
\name{print.table_design}
\alias{print.table_design}
\title{Print a table_design object}
\usage{
\method{print}{table_design}(x, ...)
}
\arguments{
\item{x}{any \R object}

\item{...}{ignored}
}
\value{
\code{x} (invisibly)
}
\description{
Print a table_design object
}
