% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_tabde.R
\name{read_tabde}
\alias{read_tabde}
\alias{read_tabde_sql}
\alias{read_tabde_fwf}
\alias{write_tabde}
\title{Read a table definition file}
\usage{
read_tabde(file)

read_tabde_sql(file)

read_tabde_fwf(file)

write_tabde(x, file, overwrite = FALSE)
}
\arguments{
\item{file}{\code{character} scalar. File that contains a table design}

\item{x}{a \link{table_design} Object}

\item{overwrite}{\code{scalar} character. Overwrite \code{file} if it exists?}
}
\value{
\code{read_tabde()} and friends return a \verb{table_design data.frame}

\code{write_tabde()} and \code{use_tabde()} return \code{file} (invisibly). This
is useful for piping the saved file, for example into
\href{https://github.com/s-fleck/shed}{\pkg{shed}} (Shiny CSV Editor).
}
\description{
tabde recommends that you store table designs as semicolon (\verb{;}) separated
csv files with header and without rownames. \code{read_tabde()} reads such files
and assigns the correct table_design subclasses based on the columns that
are present in \code{file} (f.e. \code{table_design_sql}, \code{table_design_fwf}).
}
