% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/td_fread.R
\name{td_fread}
\alias{td_fread}
\title{Read csv files fast using the data.table package and table_designs}
\usage{
td_fread(input, file, text, cmd, table_design, ...)
}
\arguments{
\item{input}{ A single character string. The value is inspected and deferred to either \code{file=} (if no \\n present), \code{text=} (if at least one \\n is present) or \code{cmd=} (if no \\n is present, at least one space is present, and it isn't a file name). Exactly one of \code{input=}, \code{file=}, \code{text=}, or \code{cmd=} should be used in the same call. }

\item{file}{ File name in working directory, path to file (passed through \code{\link[base]{path.expand}} for convenience), or a URL starting http://, file://, etc. Compressed files ending \code{.gz} and \code{.bz2} are supported if the \code{R.utils} package is installed. }

\item{text}{ The input data itself as a character vector of one or more lines, for example as returned by \code{readLines()}. }

\item{cmd}{ A shell command that pre-processes the file; e.g. \code{fread(cmd=paste("grep",word,"filename")}. See Details. }

\item{table_design}{a \link{table_design}}

\item{...}{passed on to \code{\link[data.table:fread]{data.table::fread()}}}
}
\value{
A \code{data.table} by default, otherwise a \code{data.frame} when argument \code{data.table=FALSE}.
}
\description{
A wrapper around \code{\link[data.table:fread]{data.table::fread()}} that automatically fills the
\code{col.names} and \code{colClasses} parameters from a table design.
}
