% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabde_constraints.R
\name{toString.sql_header}
\alias{toString.sql_header}
\title{Convert sql_header to a character string}
\usage{
\method{toString}{sql_header}(x, ...)
}
\arguments{
\item{x}{any \R object}

\item{...}{ignored}
}
\value{
a \code{character} scalar
}
\description{
Convert sql_header to a character string
}
