% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabde_constraints.R
\name{toString.tabde_constraints}
\alias{toString.tabde_constraints}
\title{Convert tabde_constraints to a character string}
\usage{
\method{toString}{tabde_constraints}(x, ...)
}
\arguments{
\item{x}{any \R object}

\item{...}{ignored}
}
\value{
a \code{character} scalar
}
\description{
Convert tabde_constraints to a character string
}
