% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_tabde.R
\name{use_tabde}
\alias{use_tabde}
\alias{use_tabde_fwf}
\alias{use_tabde_sql}
\title{Use table design file}
\usage{
use_tabde(
  x,
  file = {     if (assert_namespace("rprojroot"))         
    rprojroot::find_package_root_file("inst", "table_design",
    paste0(deparse(substitute(x)), ".csv"))     else stop("argument 'file' is missing") },
  overwrite = FALSE
)

use_tabde_fwf(
  x,
  file = {     if (assert_namespace("rprojroot"))         
    rprojroot::find_package_root_file("inst", "table_design",
    paste0(deparse(substitute(x)), ".csv"))     else stop("argument 'file' is missing") },
  overwrite = FALSE
)

use_tabde_sql(
  x,
  file = {     if (assert_namespace("rprojroot"))         
    rprojroot::find_package_root_file("inst", "table_design",
    paste0(deparse(substitute(x)), ".csv"))     else stop("argument 'file' is missing") },
  overwrite = FALSE
)
}
\arguments{
\item{x}{a \code{\link{table_design}} or a regular \code{data.frame}. Regular \code{data.frames}
are  converted to \code{table_designs} via \code{\link[=get_tabde]{get_tabde()}}.}

\item{file}{scalar \code{character}. name for the table_design \code{.csv} file. The
default is to construct a filename from the name of the object passed to
\code{x}.}

\item{overwrite}{\code{logical} scalar. Set \code{TRUE} to overwrite existing files.}
}
\description{
\code{use_tabde()} is designed to be used during package development. It saves
\code{table_designs} to the \verb{inst/table_design/} folder of the current
package (and creates this folder if it doesn't exist).

\code{use_tabde_fwf()} and \code{use_tabde_sql()} are convenience functions that do the
same and add the respective dummy columns.
}
