% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/values.R
\name{values}
\alias{values}
\title{Value-Domains For Columns in Table Designs}
\usage{
values(domain, value)
}
\arguments{
\item{domain}{name oder id of the domain.}

\item{value}{values that the \code{domains} can take}
}
\value{
a \code{data.frame} of class \code{"tabde_values"}
}
\description{
Value-Domains For Columns in Table Designs
}
\examples{

vals <- values(
  c("letters", "letters", "numbers"),
  c("a", "b", 2)
)

td <- tabde(
  col_name   = c("Alpha", "Num"),
  col_type   = c("character", "numeric"),
  col_domain = c("letters", "numbers")
)

df <- data.frame(
  Alpha = c("a", "b", "c"),
  Num   = c(2, 2, 3),
  stringsAsFactors = FALSE
)


# Fails because df contains values not in `vals`:
matches_tabde(df, td, values = vals)

# Ok:
matches_tabde(df[1:2, ], td, values = vals)

# domains are only checked if `values` is passed to matches_tabde
matches_tabde(df, td)
}
