% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_col_spec.R
\name{as_col_spec}
\alias{as_col_spec}
\title{Convert character data type description to \code{readr::col_spec}}
\usage{
as_col_spec(x)
}
\arguments{
\item{x}{a \code{character} vector of data types.}
}
\value{
a \code{col_spec} object (see \code{\link[readr:cols]{readr::cols()}})
}
\description{
Convert character data type description to \code{readr::col_spec}
}
\examples{

\dontrun{
as_col_spec(c("integer", "logical", "numeric"))
}

}
