% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_sql.R
\name{sql_create_table}
\alias{sql_create_table}
\title{Generate SQL CREATE TABLE statements}
\usage{
sql_create_table(
  tname,
  col_names,
  col_types,
  col_opts = rep("", length(col_names)),
  const_names = NULL,
  const_types = NULL,
  const_cols = NULL
)
}
\arguments{
\item{tname}{\code{character} scalar. Name of target sql table}

\item{col_names}{\code{character} vector. Column names of target sql table}

\item{col_types}{\code{character} scalar. Column types of target sql table.
Columns of type \code{NA} will be skipped}

\item{col_opts}{column options of target sql table (for example \verb{NOT NULL})}
}
\value{
a \verb{CREATE TABLE} statement as a \code{character} scalar
}
\description{
Creates SQL \verb{CREATE TABLE} statements from a vector of column names and
a vector of column types
}
\examples{
sql_create_table(
  "example.table",
  c("numbers", "animals"),
  c("integer", "varchar(8)"),
  c("NOT NULL", "")
)
}
