% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabde_constraints.R
\name{tabde_constraints}
\alias{tabde_constraints}
\title{Table design SQL constraints}
\usage{
tabde_constraints(const_name, const_type, const_cols, ...)
}
\arguments{
\item{const_name}{\code{character} vector. Name of the constraint.}

\item{const_type}{\code{character} vector. Type of the constraint. Currently
the only supported value is \code{"PRIMARY KEY"} but foreign keys will be
supported in the future}

\item{const_cols}{a \code{list} of \code{character} vectors that must be the same
length as \code{const_name}. If each constraint only consists of a single
column, \code{const_cols} may also be a \code{character} vector of the same length
as \code{const_name}.}

\item{...}{extra columns to be added to the resulting data.frame}
}
\value{
a \code{data.frame}
}
\description{
Table design SQL constraints
}
\examples{
tabde_constraints("PERSON_PK", "primary key", "name")
tabde_constraints("PERSON_PK", "primary key", list(c("first_name", "last_name")))
}
